+++ cloudinary-core/src/test/java/com/cloudinary/test/CloudinaryTest.java
@@ -493,6 +493,13 @@ public class CloudinaryTest {
     }
 
     @Test
+    public void testUseFetchFormat() {
+        // should support use fetch format, adds the format but not an extension
+        String result = cloudinary.url().format("jpg").useFetchFormat(true).generate("old_logo");
+        assertEquals("http://res.cloudinary.com/test123/image/upload/f_jpg/old_logo", result);
+    }
+
+    @Test
     public void testEffect() {
         // should support effect
         Transformation transformation = new Transformation().effect("sepia");
@@ -1257,7 +1264,12 @@ public class CloudinaryTest {
                 new FetchLayer().url("https://test").resourceType("image"),
                 "fetch:aHR0cHM6Ly90ZXN0",
                 new FetchLayer().url("https://test"),
-                "fetch:aHR0cHM6Ly90ZXN0"};
+                "fetch:aHR0cHM6Ly90ZXN0",
+                new FetchLayer().url("https://test").resourceType("video"),
+                "video:fetch:aHR0cHM6Ly90ZXN0",
+                new FetchLayer().url("https://www.test.com/test/JE01118-YGP900_1_lar.jpg?version=432023"),
+                "fetch:aHR0cHM6Ly93d3cudGVzdC5jb20vdGVzdC9KRTAxMTE4LVlHUDkwMF8xX2xhci5qcGc_dmVyc2lvbj00MzIwMjM="
+        };
 
         for (int i = 0; i < tests.length; i += 2) {
             Object layer = tests[i];