+++ cloudinary-http42/src/main/java/com/cloudinary/http42/ApiStrategy.java
@@ -34,9 +34,6 @@ public class ApiStrategy extends AbstractApiStrategy {
     public ApiResponse callApi(HttpMethod method, Iterable<String> uri, Map<String, ? extends Object> params, Map options) throws Exception {
         if (options == null) options = ObjectUtils.emptyMap();
 
-        String prefix = ObjectUtils.asString(options.get("upload_prefix"), ObjectUtils.asString(this.api.cloudinary.config.uploadPrefix, "https://api.cloudinary.com"));
-        String cloudName = ObjectUtils.asString(options.get("cloud_name"), this.api.cloudinary.config.cloudName);
-        if (cloudName == null) throw new IllegalArgumentException("Must supply cloud_name");
         String apiKey = ObjectUtils.asString(options.get("api_key"), this.api.cloudinary.config.apiKey);
         String apiSecret = ObjectUtils.asString(options.get("api_secret"), this.api.cloudinary.config.apiSecret);
         String oauthToken = ObjectUtils.asString(options.get("oauth_token"), this.api.cloudinary.config.oauthToken);
@@ -44,7 +41,7 @@ public class ApiStrategy extends AbstractApiStrategy {
         int timeout = ObjectUtils.asInteger(options.get("timeout"), this.api.cloudinary.config.timeout);
         validateAuthorization(apiKey, apiSecret, oauthToken);
 
-        String apiUrl = createApiUrl(uri, prefix, cloudName);
+        String apiUrl = createApiUrl(uri, options);
 
         return getApiResponse(method, params, apiKey, apiSecret, oauthToken, contentType, timeout, apiUrl);
     }