+++ cloudinary-test-common/src/main/java/com/cloudinary/test/AbstractUploaderTest.java
@@ -571,11 +571,12 @@ abstract public class AbstractUploaderTest extends MockableTest {
         assertEquals("raw", resource.get("resource_type"));
         assertTrue(resource.get("public_id").toString().startsWith("cldupload"));
 
-        resource = cloudinary.uploader().uploadLarge(new FileInputStream(temp), asMap("chunk_size", 5243000, "tags", tags));
+        resource = cloudinary.uploader().uploadLarge(new FileInputStream(temp), asMap("filename", "test123", "chunk_size", 5243000, "tags", tags));
         assertArrayEquals(tags, ((java.util.ArrayList) resource.get("tags")).toArray());
         assertEquals("image", resource.get("resource_type"));
         assertEquals(1400, resource.get("width"));
         assertEquals(1400, resource.get("height"));
+        assertEquals("test123", resource.get("original_filename"));
 
         resource = cloudinary.uploader().uploadLarge(temp, asMap("chunk_size", 5880138, "tags", tags));
         assertArrayEquals(tags, ((java.util.ArrayList) resource.get("tags")).toArray());
@@ -830,5 +831,12 @@ abstract public class AbstractUploaderTest extends MockableTest {
         Assert.assertEquals("display_name_test", uploadParams.get("display_name"));
         Assert.assertEquals(true, uploadParams.get("use_asset_folder_as_public_id_prefix"));
         Assert.assertEquals(true, uploadParams.get("visual_search"));
+    }
+
+    @Test
+    public void testNotificationUrl() {
+        Map options = asMap("notification_url", "https://www.test.com");
+        Map uploadParams = Util.buildUploadParams(options);
+        Assert.assertEquals("https://www.test.com", uploadParams.get("notification_url"));
     }
 }